requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pyaudio
google-generativeai
python-dotenv
